Reworked the last iteration of the project
Had to rework the project because it made me wanna commit toaster bath. Gameplay is incomplete but well on it's way towards completion. Rooms and Dungeon were done by Mabel, love that for her. Menu is a VIEW class that should not handle the creation of anything or any logic, strictly displays and Strings. DoWork might not be necessary. PersonFactory makes all Monsters and creates Characters. We need to only give these classes a single purpose. Menu makes Strings or displays, PersonFactory makes InteractPerson(s) / InteractPeople ? The new artifact now has a display function which can be used as opposed to sout. The attack method was made for Character, must be made for Monster now. InteractPerson, Character, and Monster are a mess, but they're my mess. Please just ask if any of the new changes are confusing.